requirements.txt
Flask>=3.1.0
numpy
pandas
psutil>=5.9.8
python-dotenv
tinytuya>=1.16.3
